Move DeletingNotification and make LocalizationService internal
